perf: convert Footer and MainBanner to server components
