+++ app/components/footer.tsx
@@ -1,8 +1,6 @@
-'use client'
-
 import Link from 'next/link'
-import { usePathname, useRouter } from 'next/navigation'
 import { profile } from 'app/lib/profile'
+import { NewsletterFooterButton } from 'app/components/newsletter-footer-button'
 
 export function ArrowIcon() {
   return (
@@ -23,9 +21,6 @@ export function ArrowIcon() {
 }
 
 export default function Footer() {
-  const pathname = usePathname()
-  const router = useRouter()
-
   return (
     <footer className="pt-8 pb-12">
       <div className="font-sm text-secondary-inv mt-8 flex items-center">
@@ -73,20 +68,7 @@ export default function Footer() {
             </a>
           </li>
           <li>
-            <button
-              onClick={() => {
-                if (pathname === '/') {
-                  window.dispatchEvent(new Event('openNewsletter'))
-                } else {
-                  sessionStorage.setItem('newsletter_highlight', '1')
-                  router.push('/')
-                }
-              }}
-              className="flex cursor-pointer items-center transition-all hover:text-neutral-800 focus-visible:text-neutral-800 focus-visible:outline-none dark:hover:text-neutral-100 dark:focus-visible:text-neutral-100"
-            >
-              <ArrowIcon />
-              <span className="ml-2">newsletter</span>
-            </button>
+            <NewsletterFooterButton />
           </li>
           <li>
             <Link
